refactor: extract incomplete mint tx funcs to file
